Add support for transactional systems to update_kernel for KOTD

+++ lib/kernel.pm
@@ -12,7 +12,8 @@ use base Exporter;
 use testapi;
 use strict;
 use utils;
-use version_utils 'is_sle';
+use version_utils qw(is_sle is_transactional);
+use transactional;
 use warnings;
 
 our @EXPORT = qw(
@@ -28,13 +29,16 @@ sub remove_kernel_packages {
     my @packages;
 
     if (check_var('SLE_PRODUCT', 'slert')) {
-        @packages = qw(kernel-rt kernel-rt-devel kernel-source-rt);
+        # workaround for bsc1227773
+        @packages = qw(kernel-rt);
+        push @packages, qw(kernel-rt-devel kernel-source-rt) unless is_transactional;
     }
     elsif (get_kernel_flavor eq 'kernel-64kb') {
         @packages = qw(kernel-64kb*);
     }
     else {
-        @packages = qw(kernel-default kernel-default-devel kernel-macros kernel-source);
+        @packages = qw(kernel-default);
+        push @packages, qw(kernel-default-devel kernel-macros kernel-source) unless is_transactional;
     }
 
     # SLE12 and SLE12SP1 has xen kernel
@@ -44,7 +48,12 @@ sub remove_kernel_packages {
 
     push @packages, "multipath-tools"
       if is_sle('>=15-SP3') and !get_var('KGRAFT');
-    zypper_call('-n rm ' . join(' ', @packages), exitcode => [0, 104]);
+
+    if (is_transactional) {
+        trup_call 'pkg remove ' . join(' ', @packages);
+    } else {
+        zypper_call('-n rm ' . join(' ', @packages), exitcode => [0, 104]);
+    }
 
     return @packages;
 }
